+++ src/pages/Pots/pages/AppliedPot/AppliedPot.style.ts
