+++ src/pages/Pots/pages/AppliedPot/components/PotInformationCard.style.ts
