+++ src/pages/Pots/pages/AppliedPot/components/PotInformationCard.tsx
