+++ src/pages/Pots/pages/AppliedPot/components/index.ts
@@ -1,5 +1,5 @@
 import PotInformationCard from "./PotInformationCard";
 
 export {
-    PotInformationCard,
-}
+  PotInformationCard
+};
