+++ src/pages/Pots/pages/MadePot/MadePot.style.ts
