+++ src/pages/Pots/pages/MadePot/components/RecruitingMyPotCard.style.ts
