+++ src/pages/Pots/pages/MadePot/components/RecruitingMyPotCard.tsx
