+++ src/pages/Pots/pages/MadePot/components/Section.style.ts
