+++ src/pages/index.ts
@@ -6,10 +6,10 @@ import SearchPage from "./Search/Search";
 import SettingPage from "./Setting/Setting";
 import SearchResultPage from "./SearchResult/SearchResult";
 import PotDetailPage from "./PotDetail/PotDetail";
-import PotMain from "./Pots/PotMain";
-import AllPotPage from "./Pots/AllPot/AllPotPage";
-import AppliedPotPage from "./Pots/AppliedPot/AppliedPot";
-import MadePotPage from "./Pots/MadePot/MadePot";
+import PotMain from "./Pots/PotMain/PotMain";
+import AllPotPage from "./Pots/pages/AllPot/AllPotPage";
+import AppliedPotPage from "./Pots/pages/AppliedPot/AppliedPot";
+import MadePotPage from "./Pots/pages/MadePot/MadePot"; 
 import SignUpPage from "./SignUp/SignUp";
 import CreatePotPage from "./CreatePot/CreatePot";
 import MyPotMainPage from "./MyPot/MyPotMain";
